inspectphoto/requirements.txt
werkzeug
flask
Pillow
extcolors
pandas
easydev
colormap